✨ api: move tests to proper bounded context
